Glob in quotes again
